tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssstyledeclaration-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSStyleDeclaration - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssstyledeclaration-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSStyleDeclaration' in window);

      }, '\'CSSStyleDeclaration\' in window');

      test(function(){

        assert_true(CSSStyleDeclaration instanceof Function);

      }, 'CSSStyleDeclaration instanceof Function');

      test(function(){

        assert_true('prototype' in CSSStyleDeclaration);

      }, '\'prototype\' in CSSStyleDeclaration');

      test(function(){

        assert_equals(CSSStyleDeclaration.prototype.toString(), '[object CSSStyleDeclarationPrototype]');

      }, 'CSSStyleDeclaration.prototype.toString() == \'[object CSSStyleDeclarationPrototype]\'');

      test(function(){

        assert_true('constructor' in CSSStyleDeclaration.prototype);

      }, '\'constructor\' in CSSStyleDeclaration.prototype');

      test(function(){

        assert_equals(CSSStyleDeclaration.prototype.constructor,CSSStyleDeclaration);

      }, 'CSSStyleDeclaration.prototype.constructor === CSSStyleDeclaration');

      // operations

      test(function(){

        assert_true('getPropertyValue' in CSSStyleDeclaration.prototype);

      }, '\'getPropertyValue\' in CSSStyleDeclaration.prototype');

      test(function(){

        assert_true(CSSStyleDeclaration.prototype.getPropertyValue instanceof Function);

      }, 'CSSStyleDeclaration.prototype.getPropertyValue instanceof Function');

      test(function(){

        assert_true('getPropertyPriority' in CSSStyleDeclaration.prototype);

      }, '\'getPropertyPriority\' in CSSStyleDeclaration.prototype');

      test(function(){

        assert_true(CSSStyleDeclaration.prototype.getPropertyPriority instanceof Function);

      }, 'CSSStyleDeclaration.prototype.getPropertyPriority instanceof Function');

      test(function(){

        assert_true('setProperty' in CSSStyleDeclaration.prototype);

      }, '\'setProperty\' in CSSStyleDeclaration.prototype');

      test(function(){

        assert_true(CSSStyleDeclaration.prototype.setProperty instanceof Function);

      }, 'CSSStyleDeclaration.prototype.setProperty instanceof Function');

      test(function(){

        assert_true('removeProperty' in CSSStyleDeclaration.prototype);

      }, '\'removeProperty\' in CSSStyleDeclaration.prototype');

      test(function(){

        assert_true(CSSStyleDeclaration.prototype.removeProperty instanceof Function);

      }, 'CSSStyleDeclaration.prototype.removeProperty instanceof Function');

    ]]>
    </script>
  </body>
</html>
